requirements/development.txt
-r base.txt

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
django-stubs==4.2.7

# Debugging
django-debug-toolbar==4.2.0
ipython==8.19.0

# Pre-commit
pre-commit==3.6.0